scaffold pledge struct
