+++ types/src/core_types/mod.rs
@@ -4,3 +4,4 @@ pub mod fixtures;
 pub mod legacy_currency_info;
 pub mod mode_ol;
 pub mod network_playlist;
+pub mod pledge;
